+++ src/components/lib/NotFound.jsx
@@ -2,11 +2,10 @@ export default function NotFound() {
 
   return (
     <>
-      <i className="bi-alarm"></i>
-      <div>Page Not Found!</div>
-      <p className="read-the-docs">
-        Are you sure you have the right URL?
-      </p>
+      <div className="w-full py-6">
+        <h1 className="text-center font-bold leading-tight my-6 text-4xl">Page Not Found</h1>
+        <p className="text-center font-light leading-tight my-6 text-black text-xl">Are you sure you have the right URL?</p>
+      </div>
     </>
   )
 }