+++ src/starter-components/SettingsView.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
-import { getState } from '../../context.jsx';
-import constants from "../../constants.json";
+import { getState } from '@/context.jsx';
+
 
 
 export default function SettingsView() {